+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/extension/ExtensibleResourceAdapter.java
@@ -106,4 +106,14 @@ public abstract class ExtensibleResourceAdapter<T> extends ResourceAdapter<T> im
     return newInstance().init(resource.fieldValidation(fieldValidation), client);
   }
 
+  @Override
+  public ExtensibleResource<T> fieldManager(String manager) {
+    return newInstance().init(resource.fieldManager(manager), client);
+  }
+
+  @Override
+  public ExtensibleResource<T> forceConflicts() {
+    return newInstance().init(resource.forceConflicts(), client);
+  }
+
 }